[stdlib] Set, Dictionary: Enable per-instance hash seeds
The performance of operations merging two similar-sized sets or dictionaries is currently quadratic. Enabling per-instance seeds makes most of these linear, which is what most people expect.

This affects Set.union, Set.symmetricDifference and similar operations, including the for-loop based forms that people write manually.

The only remaining quadratic cases are when two sets/dictionaries are merged that are mutated copy-on-write copies of the same original instance. (We can’t change the hash seed without a full rehash,